requirements.txt
asgiref==3.5.2
beautiful==0.0.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
djangorestframework==3.13.1
html5lib==1.1
idna==3.3
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.12
webencodings==0.5.1
